Fixed a corner case bug with finding the neighbors for denoising
The problem was that when there was no noise, looking at equation 19 of
https://arxiv.org/pdf/1606.05100.pdf, if a is the vector of interest,
b could be another random vector and c is a, d is b, and then the variance
is in terms of ne1 = b, which is wrong. This is when there is block matrix
no noise.

This was fixed by breaking ties by picking a vector that looks most like a